% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svdc.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(a, b)
}
\arguments{
\item{a}{The first number to add}

\item{b}{The second number to add to the first}
}
\description{
A method for adding two numbers without using a plus sign
}
\examples{
a <- 1
b <- 2
add(a,b)
}

