% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svdc.R
\name{check_file_argument}
\alias{check_file_argument}
\title{Check filename argument}
\usage{
check_file_argument(filename)
}
\arguments{
\item{filename}{The filename to check}
}
\value{
character vector with normalized path
}
\description{
Check the filename and return normalized path.
}
\keyword{internal}

