% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svdc.R
\name{data_cleaning}
\alias{data_cleaning}
\title{Data cleaning of data for use with SVAMP}
\usage{

  data_cleaning(svasss_dataset = system.file("extdata/SVASSS.alarms.data_sample.rda",
  package = "svdc"), ppn_dataset = system.file("extdata/ppn_sample.csv",
  package = "svdc"),
  movements_dataset = system.file("extdata/ani_move_sample.csv", package =
  "svdc"), svala_dataset = system.file("extdata/svala.data_sample.rda",
  package = "svdc"))
}
\arguments{
\item{svasss_dataset}{Path to dataset of SVASSS data}

\item{ppn_dataset}{Path to dataset of farm ppns}

\item{movements_dataset}{Path to animal movement dataset}

\item{svala_dataset}{Path to SVALA dataset}
}
\value{
A compound list of datasets
}
\description{
A data cleaning process needed to create data to be used in SVAMP
}

