% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{district_geo_RT90}
\alias{district_geo_RT90}
\title{Veterinary districts spatial point dataframe}
\source{
Swedish veterinary districts webpage
}
\description{
The dataset contains point locations head offices of the veterinary districts in Sweden
}

