% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_encoding.R
\name{fix_enc}
\alias{fix_enc}
\title{Convert the encoding of a dataframe from latin-1 to UTF-8}
\usage{
fix_enc(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe
}
\description{
A method for changing the encoding of a dataframe to UTF-8
}
\author{
Thomas Rosendal
}

