% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdc.R
\name{load_movement_data}
\alias{load_movement_data}
\title{Load movement data}
\usage{
load_movement_data(filename)
}
\arguments{
\item{filename}{The filename to the dataset}
}
\value{
A \code{data.frame}
}
\description{
Load and prepare movement data.
}
\keyword{internal}

