% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdc.R
\name{load_ppn_data}
\alias{load_ppn_data}
\title{Load ppn data}
\usage{
load_ppn_data(filename)
}
\arguments{
\item{filename}{The filename to the dataset}
}
\value{
A data.frame
}
\description{
Load and prepare ppn data.
}
\keyword{internal}

