% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nuts_label}
\alias{nuts_label}
\title{The name of the counties in Sweden (Län)}
\source{
Extractions from Eurostat boundaries (\url{http://epp.eurostat.ec.europa.eu/portal/page/portal/gisco_Geographical_information_maps/popups/references/administrative_units_statistical_units_1})
}
\description{
A dataframe of names of counties, and the NUTS3 code
}

