% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdc.R
\name{rt90_sweref99}
\alias{rt90_sweref99}
\title{Read in the PPN data from jordbruksverket}
\usage{
rt90_sweref99(df)
}
\arguments{
\item{df}{dataset of PPNs}
}
\value{
a dataframe of the PPNs
}
\description{
Change all projected coordinated from sweref99 to RT90
}
\keyword{internal}

