% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svdc.R
\name{data_cleaning}
\alias{data_cleaning}
\title{Data cleaning of data for use with SVAMP}
\usage{

  data_cleaning(svasss_dataset = system.file("extdata/SVASSS.alarms.data_sample.rda",
  package = "svdc"), ppn_dataset = system.file("extdata/ppn_sample.csv",
  package = "svdc"),
  movements_dataset = system.file("extdata/ani_move_sample.csv", package =
  "svdc"))
}
\arguments{
\item{svasss_dataset}{path of dataset of svass report}

\item{ppn_dataset}{Path of dataset of farm ppns}

\item{movements_dataset}{path of animal movement dataset}
}
\value{
A compound list of datasets
}
\description{
A data cleaning process needed to create data to be used in SVAMP
}

